feat(datasets): span as an example resolver
